Create a demo website with terminal styling
This is a very simple demo that loads the content of the wiki inside of the website. It uses the terminal.css lib for styling and could be indefinitely expanded by adding more content or dynamic pages.

It's built with Gatsby, which can do a whole lot more than what I'm doing with it here. I'll add some resources to the PR for further reading.